requirements.txt
hikari==2.0.0.dev118
hikari_lightbulb==2.3.3
ujson==5.7.0
